Coverage wave 8 (isolated worktrees): 95.0% all-frameworks
Non-UI 93.9% -> 94.6% (142,814 / 151,028). All-frameworks 94.4% -> 95.0%
(193,586 / 203,871). +1,090 implemented: Accelerate +270, MPS +248,
Metal +201, AudioToolbox +118, Matter +61, Charts +40, CoreMIDI +34.

Agents ran in sparse per-slug worktrees; main stayed clean until merge.
Three model API failures (AppIntents, TabularData, TelephonyMessagingKit);
telephony still landed +24 from the worktree.

+++ full/scenekit/SceneKit.swift
@@ -41,7 +41,9 @@ public let SCN_ENABLE_METAL: Int32 = 0
 public let SCN_ENABLE_OPENGL: Int32 = 0
 
 public typealias SCNActionTimingFunction = (Float) -> Float
-public typealias SCNFloat = Float
+/// Apple-oracle probe (Xcode 26.1, macOS arm64): `SCNFloat.self == CGFloat`
+/// with `MemoryLayout<SCNFloat>.size == 8`, so this is `CGFloat`, not `Float`.
+public typealias SCNFloat = CGFloat
 public typealias SCNQuaternion = SCNVector4
 public typealias SCNAnimationDidStartBlock = (SCNAnimation, any SCNAnimatable) -> Void
 public typealias SCNAnimationDidStopBlock = (SCNAnimation, any SCNAnimatable, Bool) -> Void
